Turns, Players and GameController
Armies now have a limited number of movement points per turn and belong to specific players. Only armies of the active player can be selected and commanded.
Added GameController to handle the ending/starting of turns, switching players, etc.
TMP controls:
t - end turn of the active player
m - move all armies of the active player (if they have a move path set)

+++ Assets/Scripts/Map/TileData.cs
@@ -11,7 +11,7 @@ public class TileData
     public string description {get;}
     public Texture2D texture {get;}
 
-    public float moveCost {get;}
+    public int moveCost {get;}
     public HashSet<string> pathfindingTypes {get;}
     
     public TileData(string jsonPath)
@@ -35,7 +35,7 @@ public class TileData
             pathfindingTypes.Add(pathfindingType);
         }
 
-        moveCost = (float)jObject.GetValue("moveCost");
+        moveCost = (int)jObject.GetValue("moveCost");
     }
 
     public override string ToString()
